+++ internal/repository/role_repo.go
@@ -18,8 +18,8 @@ type RoleListOption struct {
 type RoleRepository interface {
 	Create(ctx context.Context, role *model.Role) error
 	Update(ctx context.Context, role *model.Role) error
-	Delete(ctx context.Context, id uint) error
-	FindByID(ctx context.Context, id uint) (*model.Role, error)
+	Delete(ctx context.Context, id int64) error
+	FindByID(ctx context.Context, id int64) (*model.Role, error)
 	FindByName(ctx context.Context, name string) (*model.Role, error)
 	FindByCode(ctx context.Context, code string) (*model.Role, error)
 	List(ctx context.Context, opt RoleListOption) ([]*model.Role, int64, error)
@@ -41,11 +41,11 @@ func (r *roleRepository) Update(ctx context.Context, role *model.Role) error {
 	return r.db.WithContext(ctx).Save(role).Error
 }
 
-func (r *roleRepository) Delete(ctx context.Context, id uint) error {
+func (r *roleRepository) Delete(ctx context.Context, id int64) error {
 	return r.db.WithContext(ctx).Delete(&model.Role{}, id).Error
 }
 
-func (r *roleRepository) FindByID(ctx context.Context, id uint) (*model.Role, error) {
+func (r *roleRepository) FindByID(ctx context.Context, id int64) (*model.Role, error) {
 	var role model.Role
 	if err := r.db.WithContext(ctx).First(&role, id).Error; err != nil {
 		if errors.Is(err, gorm.ErrRecordNotFound) {
